+++ src/client/components/Footer/index.js
@@ -3,23 +3,11 @@
  * <!-- Footer -->
  *************************************
  */
-
+import PropTypes from "prop-types";
 import React, { Component } from 'react';
 
-
-/*-- Apply Third-party plugins --*/
-//1) Bootstrap
-import '@uixkit.react/plugins/Bootstrap/scss/bootstrap.scss';  
-
-
-//2) Fontawesome
-import '@uixkit.react/plugins/FontAwesome/scss/fontawesome.scss';  
-import '@uixkit.react/plugins/FontAwesome/scss/v4-shims.scss';  
-import '@uixkit.react/plugins/FontAwesome/scss/solid.scss';  
-import '@uixkit.react/plugins/FontAwesome/scss/brands.scss';  
-import '@uixkit.react/plugins/FontAwesome/scss/regular.scss';  
-
-
+/*-- Apply Third-party plugins (import location should be in front of "global scripts and styles") --*/
+import '@uixkit.react/plugins/_all.js';  
 
 /*-- Apply global scripts and styles --*/
 import '@uixkit.react/components/_utilities/styles/_all.scss';