+++ src/client/components/TabsAnimated/index.js
@@ -1,0 +1,8 @@
+/* 
+ *************************************
+ * <!-- Tabs Animated -->
+ *************************************
+ */
+export { default as TabsAnimated } from '@uixkit.react/components/TabsAnimated/TabsAnimated.js';
+export { default as TabList } from '@uixkit.react/components/TabsAnimated/TabList.js';
+export { default as TabPanel } from '@uixkit.react/components/TabsAnimated/TabPanel.js';