+++ src/client/components/_third-party-plugins/Videojs/js/index.js
@@ -1,3 +1,0 @@
-import videojs from './video';
-import '@videojs/http-streaming';
-export default videojs;