+++ src/client/components/_third-party-plugins/_all.js
@@ -1,0 +1,27 @@
+/*
+  Bootstrap
+  ------------- 
+*/
+import '@uixkit.react/plugins/Bootstrap/scss/bootstrap.scss';  
+import '@uixkit.react/plugins/Bootstrap/scss-rtl/bootstrap.scss';  
+
+
+/*
+  Fontawesome
+  ------------- 
+*/
+import '@uixkit.react/plugins/FontAwesome/scss/fontawesome.scss';  
+import '@uixkit.react/plugins/FontAwesome/scss/v4-shims.scss';  
+import '@uixkit.react/plugins/FontAwesome/scss/solid.scss';  
+import '@uixkit.react/plugins/FontAwesome/scss/brands.scss';  
+import '@uixkit.react/plugins/FontAwesome/scss/regular.scss';  
+
+
+/*
+  GSAP
+  ------------- 
+*/
+import { TweenMax } from '@uixkit.react/plugins/GSAP/esm/TweenMax.js';  
+
+
+