+++ src/client/components/_utilities/_all.js
@@ -1,4 +1,3 @@
-
 /**
 
 	TABLE OF CONTENTS FOR `ALL UTILITIES`
@@ -6,40 +5,8 @@
 
 */
 
-/*
- * Core Shortcut
- *
- */
 import __ from '@uixkit.react/components/_utilities/shortcut.js';
-
-
-
-/*
- * Global variables from front pages
- *
- */
-let
-	//If the file is in the root directory, you can leave it empty.
-	//If in another directory, you can write: "/blog"
-    templateUrl,
-
-	//Eg. https://uiux.cc
-	homeUrl,
-
-	//Eg. https://uiux.cc/wp-admin/admin-ajax.php
-	ajaxUrl;
-
-
-if ( typeof APP_ROOTPATH === 'undefined' ) {
-	templateUrl = '';
-	homeUrl     = '';
-	ajaxUrl     = '';
-} else {
-	templateUrl = APP_ROOTPATH.templateUrl.replace(/\/\s*$/, '' );
-	homeUrl     = APP_ROOTPATH.homeUrl.replace(/\/\s*$/, '' );
-	ajaxUrl     = APP_ROOTPATH.ajaxUrl.replace(/\/\s*$/, '' );
-}
-
+import { templateUrl, homeUrl, ajaxUrl } from '@uixkit.react/components/_utilities/global-variables.js';
 
 
 export { 
@@ -48,4 +15,5 @@ export {
 	homeUrl,
 	ajaxUrl
 };
+